+++ src/lib/deployment.ts
@@ -1,14 +1,17 @@
 // NETN Deployment element
 import { v4 as uuidv4 } from "uuid";
 import {
+  addChildElementWithValue,
   createEmptyXMLElementFromTagName,
-  getOrCreateTagElement,
+  createXMLElementWithValue,
   getTagElement,
   getTagElements,
   getTagValue,
   getValueOrUndefined,
+  removeTagValue,
   removeTagValues,
   setOrCreateTagValue,
+  xmlToString,
 } from "./domutils.js";
 
 export interface DeploymentType {
@@ -22,7 +25,7 @@ export class Deployment {
 
   constructor(element: Element) {
     this.element = element;
-    const federateElements = getTagElements(element, "Federate");
+    const federateElements = getTagElements(element, Federate.TAG_NAME);
     for (const federateElement of federateElements) {
       this.#federates.push(new Federate(federateElement));
     }
@@ -34,6 +37,7 @@ export class Deployment {
 
   set federates(federates: (FederateType | Federate)[]) {
     this.#federates.length = 0;
+    removeTagValues(this.element, Federate.TAG_NAME);
     for (const fed of federates) {
       let federateInstance =
         fed instanceof Federate ? fed : Federate.fromModel(fed);
@@ -46,6 +50,47 @@ export class Deployment {
     this.federates = [...this.#federates, fed];
   }
 
+  assignUnitToFederate(federateHandle: string, unitHandle: string) {
+    const federate = this.getFederateById(federateHandle);
+    if (!federate) return;
+    const oldFed = this.getFederateOfUnit(unitHandle);
+    if (oldFed) this.removeUnitFromFederate(oldFed.objectHandle, unitHandle);
+    federate.addUnit(unitHandle);
+  }
+
+  removeUnitFromFederate(federateHandle: string, unitHandle: string) {
+    const federate = this.getFederateById(federateHandle);
+    if (!federate) return;
+    federate.removeUnit(unitHandle);
+  }
+
+  assignEquipmentToFederate(federateHandle: string, equipmentHandle: string) {
+    const federate = this.getFederateById(federateHandle);
+    if (!federate) return;
+    const oldFed = this.getFederateOfEquipment(equipmentHandle);
+    if (oldFed)
+      this.removeEquipmentFromFederate(oldFed.objectHandle, equipmentHandle);
+    federate.addEquipmentItem(equipmentHandle);
+  }
+
+  removeEquipmentFromFederate(federateHandle: string, equipmentHandle: string) {
+    const federate = this.getFederateById(federateHandle);
+    if (!federate) return;
+    federate.addEquipmentItem(equipmentHandle);
+  }
+
+  getFederateById(objectHandle: string): Federate | undefined {
+    return this.federates.find((f) => f.objectHandle === objectHandle);
+  }
+
+  getFederateOfUnit(objectHandle: string): Federate | undefined {
+    return this.federates.find((f) => f.units.includes(objectHandle));
+  }
+
+  getFederateOfEquipment(objectHandle: string): Federate | undefined {
+    return this.federates.find((f) => f.equipment.includes(objectHandle));
+  }
+
   updateFromObject(data: Partial<DeploymentType>) {
     Object.entries(data).forEach(([key, value]) => {
       if (key in this) {
@@ -54,6 +99,12 @@ export class Deployment {
         console.warn(`Property ${key} does not exist.`);
       }
     });
+  }
+
+  toString() {
+    if (!this.element) return "";
+    const oSerializer = new XMLSerializer();
+    return oSerializer.serializeToString(this.element);
   }
 
   static fromModel(model: Partial<DeploymentType>): Deployment {
@@ -139,7 +190,10 @@ export class Federate {
     unitsEl = createEmptyXMLElementFromTagName("Units");
     this.#units = units;
     for (const unit of units) {
-      setOrCreateTagValue(unitsEl, "Unit", unit);
+      const handle = createXMLElementWithValue("ObjectHandle", unit);
+      const unitEl = createEmptyXMLElementFromTagName("Unit");
+      unitEl.appendChild(handle);
+      unitsEl.appendChild(unitEl);
     }
     this.element.appendChild(unitsEl);
   }
@@ -156,9 +210,76 @@ export class Federate {
     equipmentEl = createEmptyXMLElementFromTagName("Equipment");
     this.#equipment = equipment;
     for (const eq of equipment) {
-      setOrCreateTagValue(equipmentEl, "EquipmentItem", eq);
+      const handle = createXMLElementWithValue("ObjectHandle", eq);
+      const eqEl = createEmptyXMLElementFromTagName("EquipmentItem");
+      eqEl.appendChild(handle);
+      equipmentEl.appendChild(eqEl);
     }
     this.element.appendChild(equipmentEl);
+  }
+
+  addUnit(unitHandle: string) {
+    if (this.units.includes(unitHandle)) {
+      return console.warn(
+        `Federate ${this.name} already contains ${unitHandle}`,
+      );
+    }
+    if (!this.units || this.units.length === 0) {
+      this.units = [unitHandle];
+    } else {
+      this.units = [...this.#units, unitHandle];
+    }
+  }
+
+  removeUnit(unitHandle: string) {
+    if (!this.units.includes(unitHandle)) {
+      return console.warn(
+        `Federate ${this.name} does not contain ${unitHandle}`,
+      );
+    }
+    this.units = this.#units.filter((u) => u !== unitHandle);
+  }
+
+  addEquipmentItem(equipmentItemHandle: string) {
+    if (this.equipment.includes(equipmentItemHandle)) {
+      return console.warn(
+        `Federate ${this.name} already contains ${equipmentItemHandle}`,
+      );
+    }
+    if (!this.equipment || this.equipment.length === 0) {
+      this.equipment = [equipmentItemHandle];
+    } else {
+      this.equipment = [...this.#equipment, equipmentItemHandle];
+    }
+  }
+
+  removeEquipmentItem(equipmentHandle: string) {
+    if (!this.equipment.includes(equipmentHandle)) {
+      return console.warn(
+        `Federate ${this.name} does not contain ${equipmentHandle}`,
+      );
+    }
+    this.equipment = this.#equipment.filter((u) => u !== equipmentHandle);
+  }
+
+  addAllUnits(units: string[]) {
+    this.units = [...new Set([...this.#units, ...units])];
+  }
+
+  addAllEquipment(equipment: string[]) {
+    this.equipment = [...new Set([...this.#equipment, ...equipment])];
+  }
+
+  removeAllUnits() {
+    const removed = this.#units.splice(0);
+    this.units = [];
+    return removed;
+  }
+
+  removeAllEquipment() {
+    const removed = this.#equipment.splice(0);
+    this.equipment = [];
+    return removed;
   }
 
   updateFromObject(data: Partial<FederateType>) {